requirement.txt
pycryptodome==3.17
eyed3==0.9.7